Move celebration view out of CookNowView
