+++ Fridge to Recipe/AuthenticationView.swift
@@ -1,0 +1,23 @@
+import SwiftUI
+
+/// Simple placeholder for user authentication flows.
+struct AuthenticationView: View {
+    var body: some View {
+        VStack(spacing: 16) {
+            Text("Authentication Required")
+                .font(.title2)
+                .bold()
+            Text("Please sign in to continue.")
+                .foregroundColor(.secondary)
+            Button("Dismiss") {
+                // Placeholder action for now
+            }
+            .padding(.horizontal, 24)
+            .padding(.vertical, 12)
+            .background(Color.blue)
+            .foregroundColor(.white)
+            .cornerRadius(8)
+        }
+        .padding()
+    }
+}
